초기화 후 머지
