--- src/App.jsx
@@ -1,26 +1,0 @@
-import React, { useState } from 'react';
-
-import Page from './Page';
-
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-
-  const { count } = state;
-
-  function handleClick(value) {
-    setState({
-      count: count + value,
-    });
-  }
-
-  return (
-    <Page
-      count={count}
-      onClick={handleClick}
-    />
-  );
-}
-
-export default App;
