--- src/Buttons.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-function Buttons({ onClick }) {
-  return (
-    <p>
-      {[1, 2, 3, 4, 5].map((i) => (
-        <button type="button" key={i} onClick={() => onClick(i)}>
-          {i}
-        </button>
-      ))}
-    </p>
-  );
-}
-
-export default Buttons;
